`tools/importer-rest-api-specs`: renaming `dataapigeneratorjson` to `repository` temporarily
This'll facilitate moving the files, intentionally staging this to make merge conflicts easier?
