+++ tools/importer-rest-api-specs/components/repository/remove_service.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package dataapigeneratorjson
+package repository
 
 import (
 	"fmt"
