+++ tools/importer-rest-api-specs/components/repository/save_service.go
@@ -1,14 +1,14 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package dataapigeneratorjson
+package repository
 
 import (
 	"fmt"
 
 	"github.com/hashicorp/pandora/tools/data-api-sdk/v1/models"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/dataapigeneratorjson/helpers"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/dataapigeneratorjson/stages"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/helpers"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/stages"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/internal/logging"
 )
