+++ tools/importer-rest-api-specs/components/repository/stages/resource_ids.go
@@ -8,8 +8,8 @@ import (
 	"path/filepath"
 
 	"github.com/hashicorp/pandora/tools/data-api-sdk/v1/models"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/dataapigeneratorjson/helpers"
-	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/dataapigeneratorjson/transforms"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/helpers"
+	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/components/repository/transforms"
 	"github.com/hashicorp/pandora/tools/importer-rest-api-specs/internal/logging"
 )
