+++ tools/importer-rest-api-specs/components/repository/transforms/metadata.go
