+++ tools/importer-rest-api-specs/components/repository/transforms/resource_id.go
