+++ tools/importer-rest-api-specs/components/repository/transforms/sdk_field_object_definition.go
