+++ tools/importer-rest-api-specs/components/repository/transforms/sdk_operation.go
