+++ tools/importer-rest-api-specs/components/repository/transforms/sdk_operation_option.go
