+++ tools/importer-rest-api-specs/components/repository/transforms/service_definition.go
