+++ tools/importer-rest-api-specs/components/repository/transforms/terraform_mapping_definition.go
