+++ tools/importer-rest-api-specs/components/repository/transforms/terraform_resource_definition.go
