+++ tools/importer-rest-api-specs/components/repository/transforms/terraform_schema_field_validation.go
