+++ tools/importer-rest-api-specs/components/repository/transforms/terraform_schema_object_definition.go
